millionaire_maker/requirements.txt
Django>=3.0,<4.0
psycopg2-binary
python-decouple==3.1
django-bootstrap-v5
django-chartjs
requests
polygon-api-client
python-dotenv==1.0.1
